Don't use sh_quote in the scp protocol
Don't use shell quoting to fix filenames sent through scp. We do hoever
need to escape \n, so replace them with the cotrol sequence alike \^J,
which is the behaviour that openssh seems to use. Do this for all files
and directoris sent over the wire.

We do need to sh_quote all paths sent with the remote scp command. Add
_sh_quote to the other scp invocation.