Don't fail on the last POPD
If we had _rename set, we will get a popd when _depth == 0. This
shouldn't cause an AssertionError.

Further commands after this popd means that the server is trying to
write files above the requested path, and should be blocked (here, those
writes will happen inside the folder, which I think is good enough).