Handle "quoting" backslashes in filenames
POSIX double-quoted strings don't require backslash characters to be
escaped _unless_ they are escaping a special character.  So, for a
filename foo\bar the quoting can be "foo\bar".  But for a filename
such as foo\$bar the quoting must be "foo\\\$bar".

Prefer to avoid escaping backslashes if not necessary to be more
compatible with Windows paths.